control/controlbase: rename from control/noise.
Updates #3488
